requirements.txt
apscheduler==3.10.4
redis==5.0.8
pymongo==4.10.1
python-dotenv==1.0.0
-e ../wnpip-shared-libraray
